announce, upload , userimage api 수정

+++ src/main/java/com/example/meetup_study/announce/AnnounceServiceImpl.java
@@ -9,6 +9,9 @@ import com.example.meetup_study.room.domain.dto.RoomDto;
 import com.example.meetup_study.user.domain.User;
 import com.example.meetup_study.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -22,24 +25,21 @@ public class AnnounceServiceImpl implements AnnounceService {
 
     private final AnnounceRepository announceRepository;
     private final UserRepository userRepository;
+    private final RedisTemplate<String, String> redisTemplate;
 
     @Transactional
     @Override
     public Optional<Announce> createAnnounce(RequestAnnounceDto requestAnnounceDto) {
         Optional<User> userOpt = userRepository.findById(requestAnnounceDto.getUserId());
 
-        if(userOpt.isPresent()){
 
-            AnnounceImage announceImage = new AnnounceImage(requestAnnounceDto.getImagePath());
+        AnnounceImage announceImage = new AnnounceImage(requestAnnounceDto.getImagePath());
 
-            Announce announce = announceRepository.save(new Announce(requestAnnounceDto.getTitle(), requestAnnounceDto.getDescription(), userOpt.get(), announceImage));
+        Announce announce = announceRepository.save(new Announce(requestAnnounceDto.getTitle(), requestAnnounceDto.getDescription(), userOpt.get(), announceImage));
 
-            userOpt.get().addAnnounce(announce);
+        userOpt.get().addAnnounce(announce);
 
-            return Optional.ofNullable(announce);
-        }else{
-            throw new IllegalArgumentException("User가 없습니다.");
-        }
+        return Optional.ofNullable(announce);
     }
 
     @Override
@@ -47,22 +47,31 @@ public class AnnounceServiceImpl implements AnnounceService {
         Optional<Announce> announceOpt = announceRepository.findById(announceId);
         if(announceOpt.isPresent()) {
             Announce announce = announceOpt.get();
-//            Long viewCount = this.incrementViewCount(id);
-//            announce.changeViewCount(viewCount);
+            Long viewCount = this.incrementViewCount(announceId);
+            announce.changeViewCount(viewCount);
         }
 
         return announceOpt;
     }
 
     @Override
-    public List<AnnounceDto> getAnnounceList() {
-        List<Announce> announce = announceRepository.findAll();
+    public List<AnnounceDto> getAnnounceList(Integer page, Integer size) {
+
+        PageRequest pageRequest = PageRequest.of(page, size);
+
+        Page<Announce> announcePage = announceRepository.findAll(pageRequest);
+
+        List<Announce> announce = announcePage.getContent();
+
         List<AnnounceDto> announceDtos = announce.stream().map(r -> new AnnounceDto().convertToAnnounceDto(r)).collect(Collectors.toList());
         return announceDtos;
     }
 
     @Override
     public Optional<AnnounceDto> updateAnnounce(AnnounceDto announceDto, Long userId) {
+
+        Optional<User> userOpt = userRepository.findById(userId);
+        if (!userOpt.isPresent()) {throw new IllegalArgumentException("존재하지 않는 유저입니다.");}
 
         Optional<Announce> announceOpt = announceRepository.findById(announceDto.getId());
         if(announceOpt.isPresent() && (announceOpt.get().getUser().getId().equals(userId))) {
@@ -82,6 +91,10 @@ public class AnnounceServiceImpl implements AnnounceService {
 
     @Override
     public Optional<AnnounceDto> deleteAnnounce(Long announceId, Long userId) {
+
+        Optional<User> userOpt = userRepository.findById(userId);
+        if (!userOpt.isPresent()) {throw new IllegalArgumentException("존재하지 않는 유저입니다.");}
+
         Optional<Announce> announceOpt = announceRepository.findById(announceId);
         if(announceOpt.isPresent() && (announceOpt.get().getUser().getId().equals(userId))) {
             announceRepository.deleteById(announceId);
@@ -90,5 +103,11 @@ public class AnnounceServiceImpl implements AnnounceService {
         }else{
             throw new IllegalArgumentException("존재하지 않는 게시글이거나, 권한이 없습니다.");
         }
+    }
+
+    private Long incrementViewCount(Long announceId){
+        String key = "announce:" + announceId + ":viewCount";
+        Long count = redisTemplate.opsForValue().increment(key);
+        return count.longValue();
     }
 }
